==> 01_load_data_to_raw.ipynb <==
from datasets import load_dataset
import pandas as pd

dataset = load_dataset("Cohere/wikipedia-22-12-de-embeddings", split="train")
dataset

dataset = dataset.remove_columns("emb")
df = dataset.to_pandas()
df

df.to_pickle("./data/wikipedia-22-12-de-no-emb-raw.pkl.gz", compression="gzip")
