==> 02_clean_texts.ipynb <==
import pandas as pd
from mltb2.text import (
    has_invisible_characters,
    has_special_whitespaces,
    clean_all_invisible_chars_and_whitespaces,
)
from tqdm import tqdm

tqdm.pandas()  # activate tqdm pandas progress bar support

df = pd.read_pickle("./data/wikipedia-22-12-de-no-emb-raw.pkl.gz")
df

df["text"].progress_apply(has_invisible_characters).value_counts(dropna=False)

df["text"].progress_apply(has_special_whitespaces).value_counts(dropna=False)

df["title"].progress_apply(has_invisible_characters).value_counts(dropna=False)

df["title"].progress_apply(has_special_whitespaces).value_counts(dropna=False)

df["text"] = df["text"].progress_apply(clean_all_invisible_chars_and_whitespaces)

df["title"] = df["title"].progress_apply(clean_all_invisible_chars_and_whitespaces)

original_data_len = len(df)
original_data_len

df.drop_duplicates("text", inplace=True, keep="first")
new_data_len = len(df)
new_data_len

print("Texts droped:", original_data_len - new_data_len)

df

df.to_pickle("./data/wikipedia-22-12-de-no-emb-clean.pkl.gz", compression="gzip")
