==> 05_prepare_context.ipynb <==
import pandas as pd
from tqdm import tqdm
from mltb2.text import has_xml_tag
from uuid import uuid4

tqdm.pandas()  # activate tqdm pandas progress bar support

df = pd.read_pickle("./data/wikipedia-22-12-de-no-emb-clean-selected-merged.pkl.gz")
df.rename(columns={"text": "context", "token_count": "context_token_count"}, inplace=True)
df

df["has_xml_tag"] = df["context"].progress_apply(has_xml_tag)
df["has_xml_tag"].value_counts(dropna=False)

df.drop(df[df["has_xml_tag"] == True].index, inplace=True)
df.drop("has_xml_tag", axis=1, inplace=True)

print(len(df))
df.drop_duplicates("context", inplace=True, keep="first")
print(len(df))

## Insight

- no duplicates available

df["context_uuid"] = df["context"].progress_apply(lambda _: str(uuid4()))
df

df.to_pickle("./data/wikipedia-22-12-de-context.pkl.gz", compression="gzip")
