==> 06_data_to_db.ipynb <==
import pandas as pd
from mltb2.arangodb import ArangoImportDataManager

df = pd.read_pickle("./data/wikipedia-22-12-de-context.pkl.gz")
df

arango_import_data_manager = ArangoImportDataManager.from_config_file("./arangodb.env")
arango_import_data_manager.import_dataframe(df, "wp-dpr", create_collection=True)
