==> 09_extract_imperative_questions_to_db.ipynb <==
import gzip
import json
from mltb2.arangodb import ArangoImportDataManager

with gzip.GzipFile("./data/wikipedia-22-12-de-train-data.json.gz", "r") as f:
    train_data = json.loads(f.read().decode("utf-8"))
assert len(train_data) == 129_591

with gzip.GzipFile("./data/wikipedia-22-12-de-test-data.json.gz", "r") as f:
    test_data = json.loads(f.read().decode("utf-8"))
assert len(test_data) == 5_000

data = train_data + test_data

all_imperative_questions = []

for d in data:
    if "imperatives" in d:
        all_imperative_questions.extend(d["imperatives"])

assert len(all_imperative_questions) == 793_143

with_sie = [q for q in all_imperative_questions if "sie" in q.lower()]
len(with_sie)

with_sie_dicts = [{"imperative_formal": iq} for iq in with_sie]
with_sie_dicts[0]

arango_import_data_manager = ArangoImportDataManager.from_config_file("./config_arangodb_imperative_informal.env")
arango_import_data_manager.import_dicts(with_sie_dicts, "imp_to_informal", create_collection=True)
