==> 11_merge_informal_imperative.ipynb <==
import gzip
import json
from pprint import pprint
import pandas as pd
import jsonlines
from collections import Counter
from mltb2.text import clean_all_invisible_chars_and_whitespaces

def select_data(data: dict):
    result = {}

    llm_response = data["llm_response"]

    # ignore uncomplete response
    if llm_response["finish_reason"] != "stop":
        return None

    result["imperative_formal"] = clean_all_invisible_chars_and_whitespaces(data["imperative_formal"])
    result["imperative_informal"] = clean_all_invisible_chars_and_whitespaces(llm_response["content"])

    return result

extracted_data = []

with gzip.open("./data/imp_to_informal_backup.jsonl.gz", "r") as gzip_in:
    jsonlines_reader = jsonlines.Reader(gzip_in)
    for data in jsonlines_reader:
        selected_data = select_data(data)
        if selected_data is not None:
            extracted_data.append(selected_data)

df = pd.DataFrame(extracted_data)
df

# check if there are duplicates in imperative_formal
df[df["imperative_formal"].duplicated(keep=False)]

## Insight

- there are no suplicate imperative_formal

# remove duplicates in imperative_informal
df.drop_duplicates("imperative_informal", inplace=True, keep=False)
df

imperative_map = {k: v for k, v in zip(df["imperative_formal"], df["imperative_informal"])}

# load train data
with gzip.GzipFile("./data/wikipedia-22-12-de-train-data.json.gz", "r") as f:
    train_data = json.loads(f.read().decode('utf-8'))
assert len(train_data) == 129_591

# load test data
with gzip.GzipFile("./data/wikipedia-22-12-de-test-data.json.gz", "r") as f:
    test_data = json.loads(f.read().decode('utf-8'))
assert len(test_data) == 5_000

all_questions = []
all_imperatives = []

for d in train_data:
    if 'questions' in d:
        all_questions.extend(d['questions'])
    if 'imperatives' in d:
        all_imperatives.extend(d['imperatives'])

for d in test_data:
    if 'questions' in d:
        all_questions.extend(d['questions'])
    if 'imperatives' in d:
        all_imperatives.extend(d['imperatives'])

print("len all_questions", len(all_questions))
print("len all_imperatives", len(all_imperatives))

counter = Counter()
counter.update(all_questions)
counter.update(all_imperatives)
counter.update(df["imperative_informal"].tolist())

duplicates = set([item for item, count in counter.items() if count > 1])
len(duplicates)

def add_imperative_informal(data: list):
    for d in data:
        # handle questions
        if "questions" in d:
            questions = d["questions"]
            del d["questions"]
            questions = [q for q in questions if q not in duplicates]  # remove duplicates
            if len(questions) > 0:
                d["question"] = questions

        # handle imperatives
        if "imperatives" in d:
            imperative_formals = d["imperatives"]
            del d["imperatives"]
            imperative_formals = [i for i in imperative_formals if i not in duplicates]  # remove duplicates
            if len(imperative_formals) > 0:
                new_imperative_informals = []
                for imperative_formal in imperative_formals:
                    if imperative_formal in imperative_map:
                        new_imperative_informal = imperative_map[imperative_formal]
                        if new_imperative_informal not in duplicates:
                            new_imperative_informals.append(new_imperative_informal)
                if len(new_imperative_informals) > 0:
                    d["imperative_informal"] = new_imperative_informals
                d["imperative_formal"] = imperative_formals
        assert "question" in d or "imperative_informal" in d or "imperative_formal" in d

add_imperative_informal(train_data)
add_imperative_informal(test_data)
assert len(train_data) == 129_591
assert len(test_data) == 5_000

train_data[0]

test_data[0]

# save train data
with gzip.GzipFile("./data/wikipedia-22-12-de-train-data-v2.json.gz", "w") as f:
    f.write(json.dumps(train_data, indent=1).encode("utf-8"))

# save test data
with gzip.GzipFile("./data/wikipedia-22-12-de-test-data-v2.json.gz", "w") as f:
    f.write(json.dumps(test_data, indent=1).encode("utf-8"))
