==> 13_count_v2_data.ipynb <==
import gzip
import json

def print_data_counts(data: list):
    question_list = []
    imperative_informal_list = []
    imperative_formal_list = []

    for d in data:
        if 'question' in d:
            question_list.extend(d['question'])
        if 'imperative_informal' in d:
            imperative_informal_list.extend(d['imperative_informal'])
        if 'imperative_formal' in d:
            imperative_formal_list.extend(d['imperative_formal'])

    print("context count:", len(data))
    print("question count:", len(question_list))
    print("questions set count:", len(set(question_list)))
    print("formal count:", len(imperative_formal_list))
    print("formal set count:", len(set(imperative_formal_list)))
    print("informal count:", len(imperative_informal_list))
    print("informal set count:", len(set(imperative_informal_list)))
    print("imperative all count:", len(imperative_informal_list + imperative_formal_list))
    print("imperative set count:", len(set(imperative_informal_list + imperative_formal_list)))
    print("all count:", len(imperative_informal_list + imperative_formal_list + question_list))
    print("all set count:", len(set(imperative_informal_list + imperative_formal_list + question_list)))


# load train data
with gzip.GzipFile("./data/wikipedia-22-12-de-train-data-v2.json.gz", "r") as f:
    train_data = json.loads(f.read().decode('utf-8'))
assert len(train_data) == 129_591
print_data_counts(train_data)

# load test data
with gzip.GzipFile("./data/wikipedia-22-12-de-test-data-v2.json.gz", "r") as f:
    test_data = json.loads(f.read().decode('utf-8'))
assert len(test_data) == 5_000
print_data_counts(test_data)
